% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ott_id.R
\name{append_ott_id}
\alias{append_ott_id}
\title{Append ott id's to cube}
\usage{
append_ott_id(tree, cube, matched)
}
\arguments{
\item{tree}{An object of class 'phylo', iow a phylogenetic tree in Newick
format that was parsed by ape::read_tree()}

\item{cube}{A dataframe with for selected taxa, the number of occurrences per
taxa and per grid cell}

\item{matched}{A dataframe, returned by running the function taxonmatch() on
a phylogenetic tree, which contains the tip labels of the tree and their
corresponding gbif_id's}
}
\value{
A dataframe which consist of all the data in the original datacube,
appended with column ott_id
}
\description{
This function uses the table produced by the taxonmatch() function to create
a linking table, and then append the ott_id's as a new field to the occurence
cube.
}
\examples{
append_ott_id(tree, cube, matched)
}
