% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_faithpd.R
\name{calculate_faithpd}
\alias{calculate_faithpd}
\title{Calculation of Faith's PD}
\usage{
calculate_faithpd(tree, species, MRCA)
}
\arguments{
\item{tree}{An object of class 'phylo', iow a phylogenetic tree in Newick
format that was parsed by ape::read_tree()}

\item{species}{A character vector where each element is a species, and more
specifically, matches a tip label of the phylogenetic tree exactly}

\item{MRCA}{Node id of the taxon that represents the most recent common ancestor
of the set of species under study}
}
\value{
A string that combines "Phylogenetic diversity:" and the calculated
value
}
\description{
This function calculates Faith's PD, based on a provided list of species
and a phylogenetic tree.
}
\examples{
calculate_faithpd(tree, c("F_alessandri", "C_kawakamii", MRCA)
}
