% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_completeness.R
\name{check_completeness}
\alias{check_completeness}
\title{Check if provided phylogenetic tree is complete and covers all species in
occurence cube}
\usage{
check_completeness(mcube)
}
\arguments{
\item{mcube}{A dataframe which is returned by the function append_ott_id(),
and contains the occurence datacube with ott_id variable appended.
format that was parsed by ape::read_tree()}
}
\value{
a list - first element is the total number of species in the
occurence cube, second element is the number of species lacking in the
phylogenetic tree.
}
\description{
This function calculates which number of species in the provided occurence
cube, is not a tip label of the provided phylogenetic tree.
}
<<<<<<< HEAD
=======
\examples{
ex_data <- retrieve_example_data()
mcube<- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
check_completeness(mcube)
}
>>>>>>> 67bcd32c845d767a9ac64ed67c0bdbc8f69f1515
