% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_example_data.R
\docType{data}
\name{file_path}
\alias{file_path}
\title{Retrieve example data}
\format{
An object of class \code{character} of length 1.
}
\usage{
file_path
}
\value{
Objects tree (a phylogenetic tree of the order Fagales), cube (an
occurrence datacube, see query specifications:
https://www.gbif.org/occurrence/download/0004018-241107131044228), grid
(EEA 1km grid for study area) and pa (Natura2000 protected area polygons for
study area)
}
\description{
This function specifies the paths to the example data and reads the example
data files in and processes them so they are ready to be used in the workflow.
}
\examples{
ex_data <- retrieve_example_data()
print(ex_data$tree)
print(ex_data$cube)
print(ex_data$grid)
print(ex_data$pa)
print(ex_data$matched_nona)
}
\keyword{datasets}
