% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tip_id.R
\name{find_node_id}
\alias{find_node_id}
\title{Find corresponding tip id's of tip labels}
\usage{
find_node_id(species, my_tree)
}
\arguments{
\item{species}{A vector of species names}

\item{my_tree}{A phylogenetic tree object, read in by ape::read_tree()}
}
\value{
a vector with tip id's
}
\description{
This function finds the tip id's (numbers) that correspond to the tip labels
(species names, ott id's,..) of a phylogenetic tree.
}
