% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map_and_indicator.R
\name{generate_map_and_indicator}
\alias{generate_map_and_indicator}
\title{Mapping PD and calculating indicator}
\usage{
generate_map_and_indicator(
  pd_cube,
  grid,
  taxon = NULL,
  bbox_custom = NULL,
  cutoff = NULL
)
}
\arguments{
\item{pd_cube}{A dataframe with a variable containing grid cell codes and
a variable with calculated pd values (output of the function get_pd_cube())}

\item{grid}{An sf object with variable detailing grid cell codes and a
geometry column}

\item{taxon}{A selected higher taxon, for which the occurrence cube was
generated. Used to generate the map's title only.}

\item{bbox_custom}{Optional, numeric vector with custom bounding box
coordinates as c(xmin, xmax, ymin, ymax)}

\item{cutoff}{A variable of type numeric which determines the cut-off point
between low PD and high PD}
}
\value{
a list \code{PDindicator}, which contains one or more maps in it's first
element, and possibly one or more indicator values in it's second element
}
\description{
This function creates, for a geographic area defined by the user, a map with
the calculated PD metric for each grid cell and the location of protected
nature areas.
}
\examples{
library(dplyr)
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
pd_cube <- get_pd_cube(mcube, ex_data$tree)
PDindicator <- generate_map_and_indicator(
  pd_cube,
  ex_data$grid,
  taxon="Fagales",
  cutoff=150)
map <- PDindicator[[1]]
indicator <- PDindicator[[2]]
}
