% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd.R
\name{get_pd}
\alias{get_pd}
\title{Find MRCA for data cube and call function to calculate PD metrics}
\usage{
get_pd(tree, species, metric = "faith")
}
\arguments{
\item{tree}{A phylogenetic tree with branch lengths}

\item{species}{A character vector with species names}

\item{metric}{Name of the PD metric to be calculated}
}
\value{
Calculated PD value
}
\description{
This function determines the MRCA of all species in the datacube
and calls the function(s) to calculate PD metrics
}
\examples{
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
aggr_cube <- aggregate_cube(mcube)
PD_cube <- aggr_cube \%>\% mutate(PD = unlist(purrr::map(orig_tiplabels, ~ get_pd(ex_data$tree, unlist(.x)))))
}
