% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd_cube.R
\name{get_pd_cube}
\alias{get_pd_cube}
\title{Find MRCA for data cube and call function to calculate PD metrics}
\usage{
get_pd_cube(mcube, tree, timegroup = NULL, metric = "faith")
}
\arguments{
\item{mcube}{An occurrence data cube with matched names appended,
product of function taxonmatch()}

\item{tree}{A phylogenetic tree with branch lengths}

\item{metric}{Name of the PD metric to be calculated}

\item{species}{A character vector with species names}
}
\value{
Calculated PD value
}
\description{
This function determines the MRCA of all species in the datacube
and calls the function(s) to calculate PD metrics
}
\examples{
library(dplyr)
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
PD_cube <- get_PD_cube(mcube, tree, metric="faith")
}
