% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdindicator.R
\name{pdindicator}
\alias{pdindicator}
\title{Calculate PD indicator}
\usage{
pdindicator(cube, cutoff)
}
\arguments{
\item{cube}{An sf dataframe containing the calculated PD metrics (column name
'PD') for each grid cell with occurrences of a selected higher taxon, and the
geometries of those grid cells.}

\item{cutoff}{A variable of type numeric which determines the cut-off point
between low PD and high PD}
}
\description{
This function calculates the PD indicator, which is the percentage of high PD
cells that fall within the boundaries of a protected area. The centroid of
the grid cell is used to determine the intersection. The cut-off for low/high
PD is determined by the user (function parameter).
}
