% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmap.R
\name{pdmap}
\alias{pdmap}
\title{Mapping PD}
\usage{
pdmap(cube, taxon, xmin, xmax, ymin, ymax)
}
\arguments{
\item{cube}{An sf dataframe containing the calculated PD metrics (column name
'PD') for each grid cell with occurrences of a selected higher taxon, and the
geometries of those grid cells.}

\item{taxon}{A selected higher taxon, for which the PD was calculated}

\item{xmin}{minimum cartesian x-coordinate of map view}

\item{xmax}{maximum cartesian x-coordinate of map view}

\item{ymin}{minimum cartesian y-coordinate of map view}

\item{ymax}{maximum cartesian y-coordinate of map view}
}
\value{
A map for the selected geographic area, which visualizes the PD
value per grid cell with a colour scale, and the boundaries of protected
nature areas.
}
\description{
This function creates, for a geographic area defined by the user, a map with
the calculated PD metric for each grid cell and the location of protected
nature areas.
}
