% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cube.R
\name{aggregate_cube}
\alias{aggregate_cube}
\title{Aggregate datacube over grid cell to create new dataframe with species list per grid}
\usage{
aggregate_cube(mcube)
}
\arguments{
\item{mcube}{An occurence datacube with appended ott_id's, as produced by the append_ott_id function}
}
\value{
A dataframe with for each grid cell
}
\description{
This function aggregates a provided datacube over grid cell id, so that a new datacube
is outputted with 3 variables that contain the lists of species that are observed for
each grid cell (as speciesKeys, ott_id's and names).
}
