% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ott_id.R
\name{append_ott_id}
\alias{append_ott_id}
\title{Append ott id's to cube}
\usage{
append_ott_id(tree, cube, matched)
}
\arguments{
\item{tree}{An object of class \code{'phylo'}, a phylogenetic tree in Newick or
Nexus format that was parsed by \code{ape::read_tree()}}

\item{cube}{A dataframe with for selected taxa, the number of occurrences per
taxa and per grid cell}

\item{matched}{A dataframe, returned by running the function \code{taxonmatch()}
on a phylogenetic tree, which contains the tip labels of the tree and their
corresponding \code{gbif_id}'s}
}
\value{
A dataframe which consist of all the data in the original datacube,
appended with columns \code{ott_id}, \code{unique_name} and \code{orig_tiplabel}
}
\description{
This function uses the table produced by the \code{taxonmatch()} function to
create a linking table, and then append the \code{ott_id}'s as a new field to the
occurrence cube.
}
\examples{
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
aggr_cube <- aggregate_cube(mcube)
}
