% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_multipolygons.R
\name{convert_multipolygons}
\alias{convert_multipolygons}
\title{Convert multisurface object to multipolygon object}
\usage{
convert_multipolygons(object)
}
\arguments{
\item{object}{An object of class multisurface}
}
\value{
An object of class multipolygon
}
\description{
Convert multisurface object to multipolygon object
}
\examples{
convert_multipolygons(cube_highPD)
}
