% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map_and_indicator.R
\name{generate_map_and_indicator}
\alias{generate_map_and_indicator}
\title{Mapping PD and calculating indicator}
\usage{
generate_map_and_indicator(
  PD_cube,
  grid,
  taxon = NULL,
  bbox_custom = NULL,
  cutoff = NULL
)
}
\arguments{
\item{PD_cube}{An sf dataframe containing the calculated PD metrics (column name
'PD') for each grid cell with occurrences of a selected higher taxon, and the
geometries of those grid cells.}

\item{grid}{An sf object with variable detailing grid cell codes and a
geometry column}

\item{taxon}{A selected higher taxon, for which the occurrence cube was
generated. Used to generate the map's title only.}

\item{bbox_custom}{Optional, numeric vector with custom bounding box
coordinates as c(xmin, xmax, ymin, ymax)}

\item{cutoff}{A variable of type numeric which determines the cut-off point
between low PD and high PD}
}
\value{
a list PDindicator, which contains one or more maps in it's first element.
and possibly one or more indicator values in it's second element
}
\description{
This function creates, for a geographic area defined by the user, a map with
the calculated PD metric for each grid cell and the location of protected
nature areas.
}
\examples{
map_PD <- PD_map(PD_cube, grid, "Musceloidea", cutoff=150)

}
