% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd.R
\name{get_pd}
\alias{get_pd}
\title{Find MRCA for data cube and call function to calculate PD metrics}
\usage{
get_pd(tree, species, metric = "faith")
}
\arguments{
\item{tree}{A phylogenetic tree with branch lengths}

\item{species}{A character vector with species names}
}
\value{
Calculated PD value
}
\description{
This function determines the MRCA of all species in the datacube
and calls the function(s) to calculate PD metrics
}
\examples{
get_pd(tree, species)
}
