% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd_cube.R
\name{get_pd_cube}
\alias{get_pd_cube}
\title{Find MRCA for data cube and call function to calculate PD metrics}
\usage{
get_pd_cube(mcube, tree, timegroup = NULL, metric = "faith")
}
\arguments{
\item{mcube}{An occurrence data cube with matched names appended,
product of function \code{taxonmatch()}}

\item{tree}{A phylogenetic tree with branch lengths}

\item{timegroup}{Optional, an integer which represents the number of years
over which occurrences need to be aggregated and the PD value calculated}

\item{metric}{Name of the PD metric to be calculated}
}
\value{
Calculated PD value
}
\description{
This function determines the MRCA of all species in the datacube
and calls the function(s) to calculate PD metrics
}
\examples{
library(dplyr)
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
PD_cube <- get_pd_cube(mcube, ex_data$tree, metric="faith")
}
