% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shiny_maps.R
\name{make_shiny_maps}
\alias{make_shiny_maps}
\title{Visualizing PD maps for time periods in tabs}
\usage{
make_shiny_maps(PDindicator, plots)
}
\arguments{
\item{PDindicator}{List containing PD plots and indicators, produced by
function generate_map_and_indicator.R}

\item{plots}{A list of PD maps produced by the function
generate_map_and_indicator(), named by their time-period.}
}
\value{
An r-shiny app with PD maps in tabs
}
\description{
This function creates produces an r-shiny app that can showcase
multiple PD maps (for seperate timeperiods) in tabs
}
\examples{
library(dplyr)
ex_data <- retrieve_example_data()
mcube <- append_ott_id(ex_data$tree, ex_data$cube, ex_data$matched_nona)
mcube <- dplyr::filter(mcube, !is.na(ott_id))
PD_cube <- get_pd_cube(mcube, ex_data$tree)
PDindicator<- generate_map_and_indicator(
  PD_cube,
  ex_data$grid,
  taxon="Fagales")
plots <- PDindicator[[1]]
\dontrun{make_shiny_maps(PDindicator, plots)}
}
