% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonmatch.R
\name{taxonmatch}
\alias{taxonmatch}
\title{Taxon matching}
\usage{
taxonmatch(tree)
}
\arguments{
\item{tree}{An object of class \code{'phylo'}, a phylogenetic tree in Newick
format that was parsed by \code{ape::read_tree()}}
}
\value{
A dataframe with columns \code{ott_id} and \code{gbif_id}
}
\description{
This function matches the tip labels of a phylogenetic tree (Taxon names or
OTT id's) with corresponding GBIF id's.
}
\examples{
\dontrun{ex_data <- retrieve_example_data()
# This can take a while!
mtable <- taxonmatch(ex_data$tree)}
}
