% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonmatch.R
\name{taxonmatch}
\alias{taxonmatch}
\title{Taxon matching}
\usage{
taxonmatch(tree)
}
\arguments{
\item{tree}{An object of class 'phylo', iow a phylogenetic tree in Newick
format that was parsed by ape::read_tree()}
}
\value{
A dataframe with columns ott_id and gbif_id
}
\description{
This function matches the tip labels of a phylogenetic tree (Taxon names or
OTT id's) with corresponding GBIF id's.
}
\examples{
tree_path <- "./data/Fagales_species.nwk"
tree <- ape::read.tree(tree_path)
mtable <- taxonmatch(tree)
}
